create dedicated file
